+++ tests/testthat/test-geom_tile_theme.R
@@ -46,3 +46,14 @@ test_that("geom_tile_theme child elements inherits from theme", {
   expect_equal(gt$gp$col, c("blue", "blue"))
   expect_equal(gt$gp$lty, c(2))
 })
+
+test_that("geom_tile_theme rejects inappropriate elements", {
+  case <- substitute(geom_tile_theme(element = NULL))
+  expect_silent(eval(case))
+
+  case <- substitute(geom_tile_theme(element = element_rect()))
+  expect_silent(eval(case))
+
+  case <- substitute(geom_tile_theme(element = element_line()))
+  expect_error(eval(case), "should be of type")
+})